wipe performer data by page
